refactor: move pattern-fill bridge to its own module
